[Infineon]Add All-Clusters example app for P6 platform

+++ src/app/clusters/barrier-control-server/barrier-control-server.cpp
@@ -319,7 +319,7 @@ bool emberAfBarrierControlClusterBarrierControlGoToPercentCallback(
         state.targetPosition  = percentOpen;
         state.delayMs         = calculateDelayMs(endpoint, state.targetPosition, &state.increasing);
         emberAfBarrierControlClusterPrintln("Scheduling barrier move from %d to %d with %dms delay", state.currentPosition,
-                                            state.targetPosition, state.delayMs);
+                                            state.targetPosition, (unsigned int) state.delayMs);
         emberAfScheduleServerTick(endpoint, BarrierControl::Id, state.delayMs);
 
         if (state.currentPosition < state.targetPosition)
